agentic_web_jira/requirements.txt
langchain
langchain-community
langchain-ollama
langchain-openai
jira
beautifulsoup4
requests
googlesearch-python
streamlit
python-dotenv
